Use current GitHub Actions versions (and avoid ::set-output)
Prevent "The following action(s) uses node12 which is deprecated"
warnings by updating to the current releases of the Actions used.
(Node.js 12 is scheduled for removal from Actions runners next month.)
Use the current releases instead of `@main` too, as we don't want to
be affected by bleeding-edge bugs.

Rewrite ::set-output as a write to $GITHUB_OUTPUT instead;
the ::save-state and ::set-output commands are also deprecated.