Complete rewrite of insert_or_update_multiple_families()
The previous INSERT ... ON DUPLICATE KEY UPDATE code inserts a
new family or, if the same project+external_id entry already exists,
updates description and coded_phenotype -- as external_id is part of
the key used to locate the record, it can't be updated.

With external_id moving to a separate table, we need to write this
logic out explicitly. We search by any external id, but only insert
the primary external id for new records. (At present, this functionality
is used only by FamilyLayer.import_families(), which parses only the
primary external id.)

Add a test case exercising both inserting and updating.