FamilyTable.get_id_map_by_internal_ids() returns only primary extids
Most users of this function want a single external id (per internal id)
that they can use to populate a pedigree or for use with seqr. (The call
in ParticipantLayer.generic_individual_metadata_importer() has complex
wants but is mostly similar.) Hence, at least for now, it is simplest
to keep the 1:1 map return type and return only the primary external ids.