switch duckdb bundling method to load from CDN (#1011)
The duckdb wasm files are not big, but not exactly tiny either. They
sit at about 34 and 38MB, this is above the 32MB limit that cloudrun has
for function responses. So this PR switches to loading them from a CDN
until we can set up a better system for serving static files