allow archiving of sequencing groups via new rest and graphql endpoints
This implementation allows for simple archiving of sequencing groups.
While this is not something that needs to happen often, there are cases
where sequencing groups are replaced by other samples and are not
automatically archived. In those cases it can be necessary to manually
archive the sequencing groups and also the sample and analysis. Sample
and analysis both already allow their "active" property to be updated so
sequencing groups were the odd one out by not allowing
deactivation/archiving.