Implement updating sample IDs in gVCF files
Ideally we might use `bcftools reheader`, but that's not available on
the same image as `samtools`. In practice, that only updates the #CHROM
line too, and as this is text it's easy to do via sed and recompressing.

Sadly tabix cannot read from standard input (for trivial reasons: it opens
its input twice to detect filetype), so we operate via a FIFO instead,
which tabix only opens once because it recognises its filename extension.