Change path key in fcf outputs dict
The outputs dict in the analysis object used a "basename" key to identify
the file path. This is confusing as the resulting outputs file object
contains both "path" AND "basename" properties. basename normally
only contains the file name and extension, so use of this to denote
a file path was a poor choice from myself :/