Actions security overhaul (#1036)
* Adds zizmor fot static analysis of Github Actions

We do not scan our action workflow definitions for security flaws,
which we probably should to ensure it is safe from template injection,
shell injection, secret leaks etc.

* Overhaul Github Actions

This is aimed at improving the security of our Github Action definitions,
to avoid usage of tokens where not necessary, remove secret leaks from
checking out the repo, and ensuring we aren't vulnerable to any obvious
template injection attacks.

* Remove quotes and id on action steps

Having quotes around the id and names are superfluous and not needed.
We also do not need the id unless referencing it in later steps.

* remove more quotes

* remove more quotes

* added yaml linting

* fix yamllint config path

* remove config arg for yamllint

* remove yamllint

* remove yamllint ignore stubs

* Revert removal of quote marks in yaml

Reverting the attempted cleanup of mixed quote marks in yaml files,
so that it can be addressed in a separate PR, keeping the changes in
this branch stricly for the security overhaul.