Avoid samtools reheader's --command option (#1062)
This option passes the command string to popen(3), which invokes the command
via a shell. This means that it is impractical to adequately quote shell
metacharacters, so the sed command decayed to ...s/<SM:ID>/... and did not
match anything.

Instead run sed(1) directly without any extra shells by using a temporary
file instead of --command.