update strawberry to fix pydantic version incompatibility
A recent update to uvicorn dependency resulted in pydantic being updated
to 2.11, this update caused an incompatibility with our version of
strawberry which was referencing an internal property `is_new_type` from
within pydantic. This requirements update upgrades to the newest version
of strawberry which contains a fix for this version incompatibility.

@see https://github.com/strawberry-graphql/strawberry/issues/3807