update strawberry to fix pydantic version incompatibility (#1064)
A recent update to uvicorn dependency resulted in pydantic being updated
to 2.11, this update caused an incompatibility with our version of
strawberry which was referencing an internal property `is_new_type` from
within pydantic. This requirements update upgrades to the newest version
of strawberry (0.268.1) which contains a fix for this version incompatibility.
The fix is available since 0.262.4 but may as well update to the latest now.

@see https://github.com/strawberry-graphql/strawberry/issues/3807