Add options to export to csv/tsv formats in Billing UI (#1066)
* Add papaparse

This change adds the papaparse library to our env
and should make the exporting to csv feature a lot
simpler to implement across all pages. It also
supports reading csv files if we ever need to ingest
csv files.

* Add export to csv/tsv functionality

This change adds UI functionality to support export
of data to csv and tsv formats using the papaparse
package.

* Add types for papaparse

This change adds missing type declarations for the
papaparse library.

* Fix linting on billing pages

* Cleanup

Minor fixes, cleanups and optimisations

* minor fixes

* Added export to csv functionality

* remove filename definition

* bump deps