refactor(GraphQL-Schema): split schema types to be in individual files
In the process of adding the new endpoints noticed that the mutations
were refactored into individual GraphQLMutation classes in separate
files. Ran into the circular dependency issue with the schema endpoints
as well. Since it was going to be a crazy mess to try and add the new
endpoints to the file figured I'd give the refactor a go. Currently
working from this commit.

SET-557