Fix horiz stacked bar chart not rerendering after table/chart toggle
react refs for html elements are null when the component first renders,
becase the html element doesn't exist at that point. The bar chart would
render on first load because the parent component would load the data,
causing several re-renders of the bar chart component as the data
loaded. Any renders after the first one would work because the div that
the chart component added the svg to would exist at that point. But
after toggling to table view and back, the chart would unmount and then
remount, and because the data was already loaded there would be no
re-renders of the parent component so the ref would be null. This update
moves the chart rendering into a useEffect with the div ref as a
dependency, ensuring that the chart will always wait until the div
exists before trying to render.