split processing time by day of week chart into two charts
The processing time by day of week chart turned out to be hard to read,
as it was using the same scale for both OSS and Walk in events, we've
had way more OSS participants than walk ins so the walk in values barely
showed up on the chart. This update splits them up into two separate
charts, breaking them out into a component to reduce code duplication.