SET-438 report error message when you don t have access to a project is confusing (#1113)
Previously, when viewing a report that the viewer did not have access to, an unhelpful error message would display and the page wouldn't load properly. These changes add an alert that explains the lack of access permission and prevents further components from loading.

Additionally, the ViewerContext was modified to pass its loading and error state to child components. This allows a loading screen to be displayed when the ViewerContext has not completed initialization.