SET-597 creating an empty cohort should raise an error (#1122)
Previously when creating a cohort, it is possible for the provided criteria or template to result in an empty cohort with no sequencing group ids. This has been changed so that an error is raised in these cases rather than allowing an empty cohort, as empty cohorts are not useful.