Task/[SET-233] Improve build/test/deploy times for metamist (#1127)
This PR focuses on reducing the runtime of the test pipeline by minimising the number of containers created during execution.
Previously, a new test database container was created for each test class. However, this is expensive. This PR modifies the setup so that a single database container is created once before any tests run and is then reused across all tests. (From the current stats, this reduces the unit test runtime by ~50%).