Ensure pre-commit hooks run under Python 3.11 (#1149)
Neither pinned version of strawberry-graphql works with Python 3.14,
which pre-commit may choose by default now that it has been released.

Instead ensure that pre-commit uses the same Python version as metamist
itself does.

Also update pre-commit's pinned strawberry-graphql to be the same version
as is pinned in the main requirements.in.