[SET-190] Handle removal from cohorts
This PR addresses introduces logic to identify **active/archived/invalid** `status` of
a cohort.

**Key Changes**
With this implementation, cohort information can now be **retrieved**
along with its `status`, and the `status` of an existing cohort can also
be **updated**.

A cohort is considered **invalid** if any of the following conditions
apply:

- Any associated sample is **inactive**, or
- Any associated sequencing_group is **archived**, 

A cohort is considered **archived** if :
- The cohort itself is explicitly marked as **inactive**, or  

Otherwise, if all samples are active, all sequencing groups are
unarchived, and the cohort is marked active, the cohort is considered
**active**.

**Accessing Cohort Status**
Cohort status can be retrieved through:
- Existing GraphQL `cohorts` queries (when the `status` field is
included)

Additionally, the GraphQL cohorts query now supports conditional
retrieval based on cohort `status`.
 
**Updating Cohort Status** 
Metamist now also supports updating the `status` of an existing cohort
through:
- A new GraphQL cohort mutation